+++ app/src/main/java/com/mobiledeveloper/vktube/ui/screens/feed/models/FeedEvent.kt
@@ -1,7 +1,5 @@
 package com.mobiledeveloper.vktube.ui.screens.feed.models
 
-import com.mobiledeveloper.vktube.ui.common.cell.VideoCellModel
-
 sealed class FeedEvent {
     object ScreenShown : FeedEvent()
     object ClearAction : FeedEvent()
