+++ app/src/main/java/com/mobiledeveloper/vktube/ui/screens/feed/models/FeedState.kt
@@ -1,7 +1,6 @@
 package com.mobiledeveloper.vktube.ui.screens.feed.models
 
 import com.mobiledeveloper.vktube.data.user.StoredUser
-import com.mobiledeveloper.vktube.ui.common.cell.VideoCellModel
 import javax.annotation.concurrent.Immutable
 
 @Immutable
