+++ app/src/main/java/com/mobiledeveloper/vktube/ui/screens/subscriptions/models/SubscriptionCellModel.kt
@@ -1,0 +1,8 @@
+package com.mobiledeveloper.vktube.ui.screens.subscriptions.models
+
+data class SubscriptionCellModel(
+    val groupId: Long,
+    val groupIcon: String,
+    val groupName: String,
+    var isIgnored: Boolean = false
+)
