+++ app/src/main/java/com/mobiledeveloper/vktube/ui/screens/subscriptions/models/SubscriptionsListState.kt
@@ -1,6 +1,5 @@
 package com.mobiledeveloper.vktube.ui.screens.subscriptions.models
 
-import com.mobiledeveloper.vktube.ui.common.cell.SubscriptionCellModel
 import javax.annotation.concurrent.Immutable
 
 @Immutable
