+++ app/src/main/java/com/mobiledeveloper/vktube/ui/screens/video/VideoViewModel.kt
@@ -6,12 +6,12 @@ import com.mobiledeveloper.vktube.data.cache.InMemoryCache
 import com.mobiledeveloper.vktube.data.comments.CommentsRepository
 import com.mobiledeveloper.vktube.data.like.LikeRepository
 import com.mobiledeveloper.vktube.data.user.UserRepository
-import com.mobiledeveloper.vktube.ui.common.cell.VideoCellModel
-import com.mobiledeveloper.vktube.ui.screens.comments.CommentCellModel
-import com.mobiledeveloper.vktube.ui.screens.comments.mapToCommentCellModel
+import com.mobiledeveloper.vktube.ui.screens.feed.models.VideoCellModel
+import com.mobiledeveloper.vktube.ui.screens.feed.models.CommentCellModel
 import com.mobiledeveloper.vktube.ui.screens.video.models.VideoAction
 import com.mobiledeveloper.vktube.ui.screens.video.models.VideoEvent
 import com.mobiledeveloper.vktube.ui.screens.video.models.VideoViewState
+import com.vk.sdk.api.wall.dto.WallWallComment
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
@@ -139,6 +139,17 @@ class VideoViewModel @Inject constructor(
                 comments = currentComments
             )
         }
+    }
+
+    fun WallWallComment.mapToCommentCellModel(): CommentCellModel {
+        return CommentCellModel(
+            messageId = id,
+            userId = fromId.value,
+            text = text,
+            userName = "Unowned user",
+            dateAdded = "Unowned date",
+            avatar = ""
+        )
     }
 
     private fun showComments() {
