+++ app/src/main/java/com/mobiledeveloper/vktube/ui/screens/video/models/VideoViewState.kt
@@ -1,9 +1,8 @@
 package com.mobiledeveloper.vktube.ui.screens.video.models
 
 import com.mobiledeveloper.vktube.data.user.StoredUser
-import com.mobiledeveloper.vktube.ui.common.cell.VideoCellModel
-import com.mobiledeveloper.vktube.ui.screens.comments.CommentCellModel
-import com.vk.sdk.api.wall.dto.WallWallComment
+import com.mobiledeveloper.vktube.ui.screens.feed.models.VideoCellModel
+import com.mobiledeveloper.vktube.ui.screens.feed.models.CommentCellModel
 
 data class VideoViewState(
     val video: VideoCellModel? = null,
